refactor(main): reorganize structure to patterns.bahavioral
