+++ src/main/java/patterns/behavioral/BubbleSort.java
@@ -1,4 +1,4 @@
-package org.example;
+package patterns.behavioral;
 
 /**
  * specific strategy to sort array
