+++ src/test/java/patterns/behavioral/StrategyTest.java
@@ -1,4 +1,4 @@
-package org.example;
+package patterns.behavioral;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
